+++ frontend/svelte/src/lib/services/utils.services.ts
@@ -1,7 +1,5 @@
 import type { Identity } from "@dfinity/agent";
 import { logWithTimestamp } from "../utils/dev.utils";
-import { markdownToHTML } from "../utils/markdown.utils";
-import { sanitize } from "../utils/security.utils";
 import { getIdentity } from "./auth.services";
 
 export type QueryAndUpdateOnResponse<R> = (options: {
@@ -61,17 +59,4 @@ export const queryAndUpdate = async <R, E>({
       .catch((error) => onError?.({ certified: true, error }))
       .finally(() => (certifiedDone = true)),
   ]);
-};
-
-/**
- * Sanitize markdown text and convert it to HTML
- */
-export const markdownToSanitizedHTML = async (
-  text: string
-): Promise<string> => {
-  const [sanitizeText, convertMarkdownToHTML] = await Promise.all([
-    sanitize(),
-    markdownToHTML(),
-  ]);
-  return convertMarkdownToHTML(sanitizeText(text ?? ""));
 };
