+++ frontend/svelte/src/tests/lib/components/ui/Markdown.spec.ts
@@ -6,42 +6,41 @@ import { render, waitFor } from "@testing-library/svelte";
 import Markdown from "../../../../lib/components/ui/Markdown.svelte";
 import { mockWaiting, silentConsoleErrors } from "../../../mocks/mock.utils";
 
+let transform: (unknown) => Promise<unknown>;
+jest.mock("../../../../lib/utils/html.utils", () => ({
+  markdownToSanitizedHTML: (value) => transform(value),
+}));
+
 describe("Markdown", () => {
-  const mockMarkdownToSanitizedHTML = (
-    transform: (value: unknown) => Promise<unknown>
-  ) =>
-    jest.mock("../../../../lib/services/utils.services", () => ({
-      markdownToSanitizedHTML: (value) => transform(value),
-    }));
-
   beforeAll(silentConsoleErrors);
-
   afterAll(() => jest.clearAllMocks());
 
   it("should render html content", async () => {
-    mockMarkdownToSanitizedHTML((value) => Promise.resolve(value));
-    const { getByText } = render(Markdown, {
-      props: { text: "test" },
+    transform = (value) => Promise.resolve(value);
+    const { getByText, queryByTestId } = render(Markdown, {
+      props: { text: "test1" },
     });
-    await waitFor(() => expect(getByText("test")).not.toBeNull());
+    await waitFor(() => expect(getByText("test1")).not.toBeNull());
+    expect(queryByTestId("markdown-text")).toBeNull();
   });
 
   it("should render spinner until the text is transformed", async () => {
-    mockMarkdownToSanitizedHTML((value) => mockWaiting(0.5, value));
-    const { container, queryByText } = render(Markdown, {
-      props: { text: "test" },
+    transform = (value) => mockWaiting(0.5, value);
+    const { container, queryByText, queryByTestId } = render(Markdown, {
+      props: { text: "test2" },
     });
 
     expect(container.querySelector("svg")).toBeInTheDocument();
     expect(container.querySelector("circle")).toBeInTheDocument();
-    await waitFor(() => expect(queryByText("test")).not.toBeNull());
+    await waitFor(() => expect(queryByText("test2")).not.toBeNull());
+    expect(queryByTestId("markdown-text")).toBeNull();
     expect(container.querySelector("svg")).not.toBeInTheDocument();
   });
 
   it("should render text content on marked error", async () => {
-    mockMarkdownToSanitizedHTML(() => {
+    transform = () => {
       throw new Error("test");
-    });
+    };
     const { queryByTestId, queryByText } = render(Markdown, {
       props: { text: "text" },
     });
