+++ frontend/svelte/src/tests/lib/utils/html.utils.spec.ts
@@ -1,8 +1,10 @@
 import {
   imageToLinkRenderer,
   markdownToHTML,
+  markdownToSanitizedHTML,
+  sanitize,
   targetBlankLinkRenderer,
-} from "../../../lib/utils/markdown.utils";
+} from "../../../lib/utils/html.utils";
 
 describe("markdown.utils", () => {
   describe("targetBlankLinkRenderer", () => {
@@ -92,12 +94,12 @@ describe("markdown.utils", () => {
     });
   });
 
-  describe("markdownToHTML", () => {
+  describe("sanitize and markdown", () => {
     let renderer: unknown;
     beforeAll(() => {
       function marked(...args) {
         renderer = args[1];
-        return args[0] + "-pong";
+        return args[0] + "-markdown";
       }
       marked.Renderer = function () {
         return {};
@@ -109,10 +111,22 @@ describe("markdown.utils", () => {
         }),
         { virtual: true }
       );
+
+      jest.mock(
+        "/assets/assets/libs/purify.min.js",
+        () => ({
+          sanitize: (value: string) => value + "-sanitize",
+        }),
+        { virtual: true }
+      );
+    });
+
+    it("should call DOMPurify.sanitize", async () => {
+      expect((await sanitize())("test")).toBe("test-sanitize");
     });
 
     it("should call markedjs/marked", async () => {
-      expect((await markdownToHTML())("ping")).toBe("ping-pong");
+      expect((await markdownToHTML())("test")).toBe("test-markdown");
     });
 
     it("should call markedjs/marked with custom renderers", async () => {
@@ -122,6 +136,12 @@ describe("markdown.utils", () => {
           image: imageToLinkRenderer,
         },
       });
+    });
+
+    it("should sanitize and convert to HTML", async () => {
+      expect(await markdownToSanitizedHTML("text")).toBe(
+        "text-sanitize-markdown"
+      );
     });
   });
 });
